refactor(eval): complete eval extension promotion to public package
Moves the eval extension from internal/extensions/eval/ to the public
extensions/eval/ import path. Updates all in-tree importers
(options.go, internal/extensions/all/register.go,
internal/bootstrap/environment_tiny.go, engine_sandbox_test.go) to
the new path.

Also includes related doc updates:
- plans/2026-03-26-environment-profiles{,-impl}.md: add ConsoleWithLoad
  profile and authorizer, correct stale internal/ references
- TODO.md: mark eval promotion [x] Done, note ConsoleWithLoad design

Enables embedders (e.g. wile-goast) to import eval.Extension publicly
and compose it onto sandboxed profiles.

Prerequisite for plans/2026-03-26-environment-profiles-impl.md.
