+++ internal/bootstrap/environment_tiny.go
@@ -42,7 +42,7 @@ import (
 	"github.com/aalpar/wile/extensions/system"
 	"github.com/aalpar/wile/extensions/threads"
 	"github.com/aalpar/wile/internal/extensions/all"
-	"github.com/aalpar/wile/internal/extensions/eval"
+	"github.com/aalpar/wile/extensions/eval"
 	ioext "github.com/aalpar/wile/internal/extensions/io"
 	nsext "github.com/aalpar/wile/internal/extensions/namespace"
 	"github.com/aalpar/wile/internal/parser"
