+++ internal/extensions/all/register.go
@@ -22,7 +22,7 @@ import (
 	"github.com/aalpar/wile/extensions/process"
 	"github.com/aalpar/wile/extensions/system"
 	"github.com/aalpar/wile/extensions/threads"
-	"github.com/aalpar/wile/internal/extensions/eval"
+	"github.com/aalpar/wile/extensions/eval"
 	"github.com/aalpar/wile/internal/extensions/io"
 	nsext "github.com/aalpar/wile/internal/extensions/namespace"
 	"github.com/aalpar/wile/registry"
